requirements.txt
fastapi
uvicorn
python-dotenv
httpx
requests
gtts